pt: clean
