+++ tests/test_fancy_indexing.py
@@ -1,62 +1,92 @@
 import unittest
-from pytraj.base import *
-from pytraj import io as mdio
-from pytraj.utils import assert_almost_equal
+import pytraj as pt
 import numpy as np
+from pytraj import Trajectory
+from pytraj.testing import aa_eq
 
+class Test0(unittest.TestCase):
+    def test_0(self):
+        traj = pt.iterload("./data/md1_prod.Tc5b.x", "./data/Tc5b.top")
+        FA = traj[:]
 
-class Test(unittest.TestCase):
+        frame0 = traj[0]
+        aa_eq(frame0[traj.top("@CA")].flatten(), traj['@CA'].xyz.flatten())
+
+        # slicing with list or array
+        indices = [1, 2, 3]
+        fa = traj[indices]
+        fa2 = FA[indices]
+        fa3 = traj[range(1, 4)]
+        fa4 = FA[range(1, 4)]
+        self.assertIsInstance(fa, Trajectory)
+        # from TrajectoryIterator
+        aa_eq(fa[0].coords, traj[1].coords)
+        aa_eq(fa[1].coords, traj[2].coords)
+        # from Trajectory
+        aa_eq(fa2[1].coords, traj[2].coords)
+        aa_eq(fa2[0].coords, traj[1].coords)
+
+        # from "range"
+        aa_eq(fa3[1].coords, traj[2].coords)
+        aa_eq(fa3[0].coords, traj[1].coords)
+        aa_eq(fa4[1].coords, traj[2].coords)
+        aa_eq(fa4[0].coords, traj[1].coords)
+
+    def test_1(self):
+        # AtomMask
+        traj = pt.iterload("./data/md1_prod.Tc5b.x", "./data/Tc5b.top")
+        fa = traj[:]
+        xyz = traj.xyz[:]
+        atm = traj.top("@CA")
+        indices = atm.indices
+
+        aa_eq(fa[0, atm], fa[0][atm])
+        aa_eq(traj[0, atm], fa[0][atm])
+        aa_eq(traj[0, atm, 0], fa[0][atm, 0])
+        aa_eq(traj[0, atm, 0], xyz[0][indices][0])
+
+class Test1(unittest.TestCase):
     def test_0(self):
         # create Trajectory from Trajing_Single
-        traj = mdio.iterload("./data/md1_prod.Tc5b.x", "./data/Tc5b.top")[:]
-        #print(traj)
-        #print(traj[8][3, 0])
-        #print(traj[8][3, 0])
-        #print(traj[3, 3, 2])
-        assert_almost_equal(traj[3, 3], traj[3][3, :])
+        traj = pt.iterload("./data/md1_prod.Tc5b.x", "./data/Tc5b.top")[:]
+        aa_eq(traj[3, 3], traj[3][3, :])
         frame1 = traj[1]
-        assert_almost_equal(frame1[0], traj[1][:, :][0])
+        aa_eq(frame1[0], traj[1][:, :][0])
         assert traj[0, 0, 0] == -16.492
-        #print(traj[0][:][-1])
-        #print(traj[0:2][0])
-        #print(traj[:, 0, 0])
-        #print(traj[:, 3].size)
-        #print(traj[:, :, 0][0, 0])
         assert traj[:, :, 0][0, 0] == traj[0, 0, 0]
         f0 = traj[0]
-        #print(f0[:, :][-10])
         farr0 = traj[:2]
-        # #print farr0[0:1, :]
-        #print("XYYYYYY")
-        #print(farr0[0:1])
 
         #self.assertRaises(NotImplementedError, lambda: traj[2:4, :, : ])
         fa = traj[2:4]
-        #print(fa[0, :][0])
-        #print(fa[0:2, :].__len__())
-        #print(fa[0:1, :].__len__())
-        #print(type(fa[0:1, :]))
-
-        #print(type(traj[:, :, :][0]))
-        #print(traj[:, :, :].__len__())
 
         # we don't support traj[:, idx] or traj[:, idx, idy] since this give wrong answer
         #  got ~0.0 value
-        #print("assert_almost_equal(traj[:, 0, 0], np.asarray(traj[0][0]))")
-        #print(traj[:, 0, 0])
-        #print(traj[0][0])
-        assert_almost_equal(traj[:, 0, 0], np.asarray(traj[0][0]))
+        aa_eq(traj[:, 0, 0], np.asarray(traj[0][0]))
 
         for i in range(traj[0].buffer2d.shape[0]):
-            ##print ("coords for atom %s" % i)
-            ##print (traj[:, :, 0][i])
-            ##print (traj[0].buffer2d[i])
-            assert_almost_equal(traj[:, :, 0][i], traj[0].buffer2d[i])
+            aa_eq(traj[:, :, 0][i], traj[0].buffer2d[i])
 
         # slicing with mask
         atm = traj.top("@CA")
         traj[atm]
         traj[:, atm]
+
+class TestSegmentationFault(unittest.TestCase):
+    def test_0(self):
+        # NOTE: no assert, just check for segfault
+        traj = pt.load("./data/md1_prod.Tc5b.x", "./data/Tc5b.top")
+        trajiter = pt.load("./data/md1_prod.Tc5b.x", "./data/Tc5b.top")
+        atm = traj.top("@CA")
+        f0 = traj[5]
+        f0 = traj[0]
+        f0.top = traj.top
+        f0['@CA']
+        traj[0, '@CA']
+
+        f0 = traj[0, '@CA']
+        f1 = traj['@CA'][0]
+        assert pt.tools.rmsd(f0.xyz, f1.xyz) == 0.0
 
 
 if __name__ == "__main__":
